+++ src/OpenTelemetry/CHANGELOG.md
@@ -10,6 +10,10 @@ Notes](../../RELEASENOTES.md).
   variables.
   ([#7187](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7187))
 
+* **Breaking Change** Explicit histogram boundaries no longer allow more than
+  10 million values.
+  ([#7165](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7165))
+
 ## 1.15.3
 
 Released 2026-Apr-21